Refactor Storage, Utils, Parser, TaskList into its own package
