+++ src/main/java/duke/utils/Parser.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.utils;
 
 
 import duke.command.ByeCommand;
