+++ src/main/java/duke/utils/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.utils;
 
 import duke.task.Task;
